Ajusta tipos e tamanhos dos campos das migrations

+++ database/migrations/2020_08_22_021043_create_transactions_table.php
@@ -10,10 +10,10 @@ class CreateTransactionsTable extends Migration
     {
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
-            $table->integer('payer_wallet_id');
-            $table->integer('payee_wallet_id');
+            $table->unsignedBigInteger('payer_wallet_id');
+            $table->unsignedBigInteger('payee_wallet_id');
             $table->decimal('ammount');
-            $table->integer('transaction_status_id');
+            $table->unsignedBigInteger('transaction_status_id');
             $table->timestamps();
         });
     }
